+++ docs/source/CHANGELOG.rst
@@ -27,6 +27,9 @@ Changed
 * Remove load_data_for_figure from plotting.utils
 * Remove CoincidentLinePlot (Same (at least similar) feature can be done with fill_between)
 * Remove zero_nans arg in shift_by. Use da.fillna(0), instead.
+* Remove cut_nan_coords. Use dropna like that:
+  for cname in da.coords:
+    da = da.dropna(dim=cname, how="any")
 
 * Change method name: S.swap_angle_unit -> S.switch_angle_unit